siteinfo.bbclass: Add powerpc-linux-gnuspe for PowerPC e500 cores.
This adds support for the PowerPC e500 cores by adding the
target tripple to siteinfo.bbclass.

Also, as openembedded-core is not yet distro-less (work in progress)
I added support for the poky.conf to configure the target tripple
correctly.

Taken from OpenEmbedded where it was run tested. Build tested here.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>